Todo app replaced with the context api
